Added some brief checkpoint code, and it works from level 1 to 2.
GameManager keeps track of the last checkpoint position since it is already a singleton who is not killed on reset of the level.
LevelTriggerGravity will get a reference to the update manager everytime it is executed and add the last position of the checkpoint.
On level restart, the player object will update its position with the data held by the GameManager.


Signed-off-by: cdavi014 <cdavi014@fiu.edu>